requirements.txt
streamlit
pandas
numpy
scikit-learn
shap
lime
matplotlib

SQLAlchemy
python-dotenv

openai
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai

tabulate